refactor: Add getById query to members module
This commit adds a new query function, getById, to the members module. The getById query retrieves a specific member by their ID and populates the associated user information. This functionality is useful for fetching member details for a given ID.

The getById query is implemented with a handler function that first checks the authentication status of the user. If the user is not authenticated, null is returned. If the member with the specified ID does not exist, null is also returned. The handler then queries the database to fetch the member and checks if the current user is a member of the same workspace. If the current user is not a member of the workspace, null is returned. Finally, the user associated with the member is populated and returned along with the member details.

This commit enhances the functionality of the members module and improves the retrieval of member information.